requirements.txt
# Core Framework
Django==5.0.6

# API and Authentication
djangorestframework==3.15.1
django-cors-headers==4.4.0

# Database
psycopg2-binary==2.9.9

# Production and Deployment
whitenoise==6.6.0

#yarn install (in terminal)